Improved address strings parsers.
Now they are also able to work without a SetupPane instance.
The idea here is to add a better way to create/parsed them, remove the customization from the setup page and move it inside a separated stateless stage which can be used in different part of the project (For quick properties for example)

+++ src/main/java/parozzz/github/com/hmi/comm/modbustcp/intermediate/doubleword/ModbusTCPReadDWordIntermediate.java
@@ -22,7 +22,7 @@ public class ModbusTCPReadDWordIntermediate extends ModbusTCPReadNumberIntermedi
     {
         //A Double Word is 4 bytes, so an integer contains it all :)
         //Every new value added, it is added to the actual value by shifting the set value by 16 * (values received)
-        this.value |= (value & 0xFFFF) << (16 * valueCounter ++);
+        this.value |= (long) (value & 0xFFFF) << (16 * valueCounter ++);
 
         if(signed && valueCounter == 2)
         {
